+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `espg::Id<T>` can now be passed by reference to `EventStore::load` to keep ownership at call site
 - `uuid` feature flag
 - `espg::Id<T>` supports `From<&str>`, `From<&String>`, `From<String>` and `From<uuid::Uuid>` (behind `uuid` feature flag)
+- `EventStore::all<X: Aggregate>()` to load all aggregates
+- `EventStore::create<X: Aggregate>(event: X::Event) -> Id<X>` to create an aggregate (behind `uuid` feature flag)
 
 ### Changed